+++ src/entities/error/api/index.ts
@@ -1,0 +1,5 @@
+export {fetchErrorById} from './fetchErrorById';
+export {fetchErrorGroupById} from './fetchErrorGroupById';
+export {fetchErrorGroups} from './fetchErrorGroups';
+export {fetchErrors} from './fetchErrors';
+export {type ErrorsStatsResponse, fetchErrorsStats} from './fetchErrorsStats';
