+++ src/entities/error/index.ts
@@ -1,0 +1,2 @@
+export * from './api';
+export type * from './model';
