+++ src/entities/error/model/Err.ts
@@ -17,13 +17,3 @@ export interface Err {
     env?: Record<string, unknown>;
     time: number;
 }
-
-export interface ErrGroup {
-    id: string;
-    message: string;
-    file: string;
-    line: number;
-    firstSeenAt: number;
-    lastSeenAt: number;
-    counter: number;
-}
