+++ src/entities/error/model/ErrrGroup.ts
@@ -1,0 +1,9 @@
+export interface ErrGroup {
+    id: string;
+    message: string;
+    file: string;
+    line: number;
+    firstSeenAt: number;
+    lastSeenAt: number;
+    counter: number;
+}
