+++ src/entities/error/model/index.ts
@@ -1,0 +1,2 @@
+export type {Err} from './Err';
+export type {ErrGroup} from './ErrrGroup';
