+++ src/features/errors/ui/ErrorGroupsTable/ErrorGroupsTable.tsx
@@ -1,7 +1,7 @@
 import {DataFetchError} from '@/shared/ui/DataFetchError';
 import {DataLoader} from '@/shared/ui/DataLoader';
 import {NavigateFunction} from 'react-router';
-import {ErrGroup} from '@/entities/error/model/types';
+import {ErrGroup} from '@/entities/error';
 import {getErrorGroupsTableColumns} from '@/features/errors/ui/ErrorGroupsTable/lib/getErrorGroupsTableColumns';
 import TableSelection from '@/shared/ui/Table/TableSelection';
 import {useState} from 'react';
